Created a separate file for top stories
